move test code to separate subpackage
mock_client.go imports testing libraries which previously were compiled
into any program importing this package. This commit moves the mock
client and server to a separate package to remove this dependency.
